requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imutils==0.5.2
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
Pillow==6.0.0
progressbar2==3.42.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-utils==2.3.0
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
